Crate ndi

NewTek NDI®1 Bindings for rust

NDI® is a registered trademark of NewTek, Inc. http://ndi.tv/ ↩

Modules

error

The error type used in this crate

find

The Find struct and related constructs for finding NDI sources

recv

The Recv struct and related constructs for receiving NDI

send

The Send struct and related constructs for sending NDI

Structs

AudioData

An audio frame

MetaData

The data description for metadata

Source

A descriptor of a NDI source available on the network.

Tally

Tally information

VideoData

Describes a video frame

Enums

FourCCAudioType

The FourCC type of a AudioData frame

FourCCVideoType

The FourCC type of a VideoData frame

FrameFormatType

A description of the frome format of a frame

FrameType

A description of the type of of frame received.

Functions

cleanup⚠

Destroy everything associated with the library

initialize

Start the library

is_supported_CPU

Recover whether the current CPU in the system is capable of running NDILib.