ndi
Function initialize

pub fn initialize() -> Result<(), NotSupported>

Start the library

This is not actually required, but will start the libraries which might get you slightly better performance in some cases. In general it is more “correct” to call this although it is not required. There is no way to call this that would have an adverse impact on anything. This will return Err if the CPU is not sufficiently capable to run NDILib currently NDILib requires SSE4.2 instructions (see documentation). You can verify a specific CPU against the library with a call to is_supported_CPU()