ndi::recv
Enum RecvColorFormat

pub enum RecvColorFormat {
    BGRX_BGRA = 0,
    UYVY_BGRA = 1,
    RGBX_RGBA = 2,
    UYVY_RGBA = 3,
    Fastest = 100,
    Best = 101,
}

Determines what color formats you are passed when a frame is received.

In general, there are two color formats used in any scenario: one that exists when the source has an alpha channel, and another when it does not. See FourCCVideoType for details on individual FourCC types.

Variants

BGRX_BGRA = 0

BGRX or BGRA

UYVY_BGRA = 1

UYVY or BGRA

RGBX_RGBA = 2

RGBX or RGBA

UYVY_RGBA = 3

UYVY or RGBA

Fastest = 100

Normally UYVY, see RecvColorFormat

If you specify the color option RecvColorFormat::Fastest, the SDK will provide buffers in the format that it processes internally without performing any conversions before they are passed to you. The allow_video_fields option is assumed to be true in this mode.

Best = 101

Varies, see RecvColorFormat

If you specify the color option RecvColorFormat::Best, the SDK will provide you buffers in the format closest to the native precision of the video codec being used. In many cases this is both high-performance and high-quality and results in the best quality. Like RecvColorFormat::Fastest this format will always deliver individual fields, implicitly assuming the allow_video_fields option as true.