ndi::recv
Struct Recv

pub struct Recv {
    pub connected: bool,
    /* private fields */
}

The NDI receiver struct

Fields
connected: bool

whether the Recv is currently connected

Implementations

impl Recv

pub fn new() -> Result<Self, RecvCreateError>

Create new receiver which isn’t connected to any sources

It is recommended that you use RecvBuilder instead if possible

pub fn connect(&mut self, source: &Source)

Connect to a source

pub fn disconnect(&mut self)

Disconnect from all sources

pub fn capture_all( &self, video_data: &mut Option<VideoData>, audio_data: &mut Option<AudioData>, meta_data: &mut Option<MetaData>, timeout_ms: u32, ) -> FrameType

Receive video, audio and metadata frames.

This call can be called simultaneously on separate threads, so it is entirely possible to receive audio, video, metadata all on separate threads. This function will return FrameType::None if no data is received within the specified timeout and FrameType::ErrorFrame if the connection is lost. Buffers captured with this must be freed with the appropriate free function.

pub fn capture_video( &self, video_data: &mut Option<VideoData>, timeout_ms: u32, ) -> FrameType

Receive video frame

pub fn capture_audio( &self, audio_data: &mut Option<AudioData>, timeout_ms: u32, ) -> FrameType

Receive audio frame

pub fn capture_metadata( &self, meta_data: &mut Option<MetaData>, timeout_ms: u32, ) -> FrameType

Receive metadata frame

pub fn get_performance(&self) -> (RecvPerformance, RecvPerformance)

Get the performance metrics (total, dropped)

pub fn get_queue(&self) -> RecvQueueSize

Get the current queue sizes

pub fn get_no_connections(&self) -> u32

Get the current number of sources connected to

pub fn set_tally(&mut self, tally: Tally)

Set tally info for sender

pub fn add_connection_metadata(&self, metadata: &MetaData)

Add a connection metadata string to the list of what is sent on each new connection.

If someone is already connected then this string will be sent to them immediately. Connection based metadata is data that is sent automatically each time a new connection is received. To reset them you need to clear them all and set them up again using Recv::recv_clear_connection_metadata()

pub fn send_metadata(&self, metadata: &MetaData) -> bool

Send metadata to sender

This returns false if we are not currently connected to anything.

pub fn recv_clear_connection_metadata(&self)

Clear all connection metadata